Disable Bash wildcard expansion (globbing)
We always want the wildcards to be expanded by Git or other
subprocesses.  Otherwise, if the shell expands them, then

FILES='./*'
git grep {regex} -- $FILES

won't necessarily check all of the files in the repository.  We could
quote the various FILES variables, but that is error-prone.  It's easier
to just disable globbing, since we don't need the shell to do that for
us in this case.

This commit also makes the necessary changes to the checkstyle script
to handle files that were previous hidden from the searches.